requirements.txt
affine==2.4.0
anywidget==0.9.18
asttokens==3.0.0
attrs==25.3.0
blessings==1.7
blinker==1.9.0
bqplot==0.12.45
branca==0.8.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
click-plugins==1.1.1.2
cligj==0.7.2
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
earthengine-api==1.6.8
ee==0.2
eerepr==0.1.2
executing==2.2.1
filelock==3.19.1
Flask==3.1.2
folium==0.20.0
fonttools==4.60.0
fsspec==2025.9.0
future==1.0.0
geemap==0.36.3
geocoder==1.38.1
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
httplib2==0.31.0
idna==3.10
ipyevents==2.0.4
ipyfilechooser==0.6.0
ipyleaflet==0.20.0
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter-leaflet==0.20.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==2.5.0
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
plotly==6.3.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.1
psygnal==0.14.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.2.4
pyperclip==1.10.0
pyshp==2.3.1
python-box==7.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
rasterio==1.4.3
ratelim==0.1.6
requests==2.32.5
rsa==4.9.1
scooby==0.10.1
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
torch==2.8.0
traitlets==5.14.3
traittypes==0.2.1
triton==3.4.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xyzservices==2025.4.0
